+++ dat/RecipesScraper/README.md
@@ -1,0 +1,18 @@
+# Scrapers
+
+This folder contains various scrapers.
+
+Scrapers
+* [All Recipes](http://allrecipes.com/)
+
+# All Recipes Scraper
+
+This scraper takes in the sitemap list of recipe hubs and scrapes all of the recipes from each one.
+
+Start the scraper:
+
+```bash
+nohup scrapy crawl allrecipes &
+```
+
+This will create a JSON file `recipes.json` in the current directory.
