+++ dat/RecipesScraper/scrapy.cfg
@@ -1,0 +1,11 @@
+# Automatically created by: scrapy startproject
+#
+# For more information about the [deploy] section see:
+# https://scrapyd.readthedocs.org/en/latest/deploy.html
+
+[settings]
+default = RecipesScraper.settings
+
+[deploy]
+#url = http://localhost:6800/
+project = RecipesScraper
