--- dat/scrape/AllRecipesScraper/AllRecipesScraper/items.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-import scrapy
-
-
-class RecipeItem(scrapy.Item):
-    """Info from recipe to be collected."""
-    recipe = scrapy.Field()
-    ingredients = scrapy.Field()
-    tags = scrapy.Field()
