--- dat/scrape/AllRecipesScraper/scrapy.cfg
@@ -1,11 +1,0 @@
-# Automatically created by: scrapy startproject
-#
-# For more information about the [deploy] section see:
-# https://scrapyd.readthedocs.org/en/latest/deploy.html
-
-[settings]
-default = AllRecipesScraper.settings
-
-[deploy]
-#url = http://localhost:6800/
-project = AllRecipesScraper
