--- dat/scrape/README.md
@@ -1,6 +1,0 @@
-# Scrapers
-
-This folder contains various scrapers.
-
-Scrapers
-* [All Recipes](http://allrecipes.com/)
